Update (11/30/2019)
This update ensures compatibility with Unity 2019.2.14f1 and the generation quite a bit.

Also new is the new random name generation method which you can supply with a format string and it outputs a random weapon name for example:

"LD m/DDDD"

It will replace all the "L"s with letters and all the "D"s with digits and leave everything else as-is.

The WeaponBody class has been removed because it was redundant.

The structure of the generated weapon game objects has been changed to make it cleaner, so instead of:

- Weapon Name
-- Weapon Body (Clone)
--- Stock Transform
---- Stock (Clone)
--- Grip Transform
---- Grip (Clone)
--- Magazine Transform
---- Magazine (Clone)
--- Barrel Transform
---- Barrel (Clone)
--- Scope Transform
---- Scope (Clone)

It's

- Weapon Name
-- Stock Transform
--- Stock
-- Grip Transform
--- Grip
-- Magazine Transform
--- Magazine
-- Barrel Transform
--- Barrel
-- Scope Transform
--- Scope

Also, "Handles" have been renamed to "Grips".